refactor: Use a separate ViewModel for preferences/settings
#222
